SDP-760 - address PR comments
